fix: corrige inconsistências no CRUD de usuário

+++ src/main/java/mythosforge/fable_minds/service/UserService.java
@@ -2,9 +2,11 @@ package mythosforge.fable_minds.service;
 
 
 
+import mythosforge.fable_minds.config.security.auhentication.dto.UpdateUserDTO;
 import mythosforge.fable_minds.models.Users;
 import mythosforge.fable_minds.repository.UserRepository;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -14,7 +16,9 @@ import java.util.Optional;
 @Service
 public class UserService implements IUserService {
 
+        @Autowired
         private final UserRepository userRepository;
+        @Autowired
         private final PasswordEncoder passwordEncoder;
 
         public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
@@ -24,7 +28,13 @@ public class UserService implements IUserService {
 
         @Override
         public Users createUser(Users user) {
-            user.setPassword(passwordEncoder.encode(user.getPassword())); // Codifique a senha
+            if (userRepository.existsByUsername(user.getUsername())) {
+                throw new RuntimeException("Nome de usuário já está em uso");
+            }
+            if (userRepository.existsByEmail(user.getEmail())) {
+                throw new RuntimeException("E-mail já está em uso");
+            }
+            user.setPassword(passwordEncoder.encode(user.getPassword())); // Codifica a senha
             return userRepository.save(user);
         }
 
@@ -39,17 +49,24 @@ public class UserService implements IUserService {
         }
 
         @Override
-        public Users updateUser(Long id, Users userDetails) {
+        public Users updateUser(Long id, UpdateUserDTO userDTO) {
             return userRepository.findById(id).map(user -> {
-                user.setUsername(userDetails.getUsername());
-                user.setEmail(userDetails.getEmail());
-                user.setPassword(userDetails.getPassword() != null ? passwordEncoder.encode(userDetails.getPassword()) : user.getPassword()); // Atualiza a senha se fornecida
+                user.setUsername(userDTO.getUsername());
+                user.setEmail(userDTO.getEmail());
+
+                if (userDTO.getPassword() != null && !userDTO.getPassword().isBlank()) {
+                    user.setPassword(passwordEncoder.encode(userDTO.getPassword()));
+                }
+
                 return userRepository.save(user);
             }).orElseThrow(() -> new RuntimeException("Usuário não encontrado com id: " + id));
         }
 
         @Override
         public void deleteUser(Long id) {
+            if (!userRepository.existsById(id)) {
+                throw new RuntimeException("Usuário com ID " + id + " não encontrado.");
+            }
             userRepository.deleteById(id);
         }
     }
